backend/requirements.txt
# Web Framework (Pydantic v2 compatible)
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
SQLAlchemy>=2.0,<3.0

# RAG stack: LangChain + Groq + Chroma + FastEmbed (Py 3.11 compatible)
# Align on langchain-core >=0.2 due to langchain-groq
langchain>=0.2.10,<0.3
langchain-community>=0.2.12,<0.3
langchain-groq>=0.1.10,<0.2
chromadb>=0.5.3,<0.6

# Embeddings (free, CPU)
fastembed>=0.5.1,<0.6

# File loaders
pypdf>=4.3.1,<5.0

# Optional/Utilities
python-dotenv>=1.0,<2
pydantic>=2.8,<3
requests>=2.32,<3

# Additional deps used by app.main -> training_service/vectorstore
# (These are required because app.main registers routes using training_service/vectorstore)
pymupdf>=1.24.2,<1.25          # provides `fitz`
sentence-transformers>=2.6,<3
faiss-cpu>=1.8.0,<1.9          # CPU FAISS; on Windows, install via conda if pip wheel unavailable